Move PublicAssetsMemoryModule to apps/memory module
