+++ tmail-backend/apps/memory/src/main/scala/com/linagora/tmail/james/app/modules/PublicAssetsMemoryModule.scala
@@ -1,6 +1,7 @@
-package com.linagora.tmail.james.jmap.publicAsset
+package com.linagora.tmail.james.app.modules
 
 import com.google.inject.{AbstractModule, Scopes}
+import com.linagora.tmail.james.jmap.publicAsset.{MemoryPublicAssetRepository, PublicAssetRepository}
 
 class PublicAssetsMemoryModule extends AbstractModule {
   override def configure(): Unit = {
